Tweak figures.

+++ Projects/mpm_mac/2017-jcp-apic-scripts/run_conv_taylor.sh
@@ -4,19 +4,23 @@ NAME=conv-taylor
 
 ARGS="../mpm_mac 16 -last_frame 1 -frame_dt 1 -mu 2 -scale_mass 3"
 
+FULL=1 # Set to 1 for a full rebuild; 0 to skip rerunning the simulations
+
 LO=32
 HI=256
 SKIP=16
-rm -rf $NAME
-mkdir -p $NAME
 
-for r in `seq $HI -$SKIP $LO` ; do
-    dt=`perl -e "print (1.0/$r)"`
-    DT="-max_dt $dt -min_dt $dt"
-    echo $ARGS $DT -no_affine -o $NAME/pic-$r -resolution $r
-    echo $ARGS $DT -affine -o $NAME/apic-$r -resolution $r
-    echo $ARGS $DT -no_affine -flip 1 -o $NAME/flip-$r -resolution $r
-done | xargs -P 16 -n 1 -d '\n' bash -c
+if [ "X$FULL" = "X1" ] ; then
+    rm -rf $NAME
+    mkdir -p $NAME
+    for r in `seq $HI -$SKIP $LO` ; do
+        dt=`perl -e "print (1.0/$r)"`
+        DT="-max_dt $dt -min_dt $dt"
+        echo $ARGS $DT -no_affine -o $NAME/pic-$r -resolution $r
+        echo $ARGS $DT -affine -o $NAME/apic-$r -resolution $r
+        echo $ARGS $DT -no_affine -flip 1 -o $NAME/flip-$r -resolution $r
+    done | xargs -P 16 -n 1 -d '\n' bash -c
+fi
 
 for t in grid particle ; do
     for s in pic apic flip ; do
@@ -29,6 +33,7 @@ for t in grid particle ; do
         sed -i -e 's/ .*L-inf: / /' -e 's/ L-2://' -e 's/ N: .*//' $NAME/$t-$s.txt
     done
 done
+
 sed -e 's/xxx/pic/' -e 's/XXX/PIC/' conv_taylor_plot.tex  > $NAME/plot-pic.tex 
 sed -e 's/xxx/apic/' -e 's/XXX/APIC/' conv_taylor_plot.tex  > $NAME/plot-apic.tex 
 sed -e 's/xxx/flip/' -e 's/XXX/FLIP/' conv_taylor_plot.tex  > $NAME/plot-flip.tex 
@@ -41,6 +46,18 @@ for t in grid particle ; do
     done
 done
 
-for s in pic apic flip ; do
-    pdflatex $NAME/plot-$s.tex 
-done
+cat <<EOF > $NAME/SConstruct
+import os
+import re
+env=Environment(ENV = os.environ)
+env['PSSUFFIX']=".eps"
+r=re.compile(".*\.tex$")
+for f in [x for x in os.listdir(".") if r.match(x)]:
+    t=env.DVI(f)
+    env.PostScript(t)
+    env.PDF(t)
+EOF
+(
+    cd $NAME
+    scons
+)
